refactor: former and main page into barrell
