+++ src/pages/former/index.ts
@@ -1,0 +1,1 @@
+export { FormerPage } from './former-page';
