+++ src/pages/index.ts
@@ -1,1 +1,3 @@
 export * from './about';
+export * from './former';
+export * from './main';
