+++ src/pages/main/main-page.test.tsx
@@ -1,12 +1,12 @@
 import { act, fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { AppContextProvider } from 'providers/app-context-provider';
 import React from 'react';
-import { CardProviderStore } from '../providers/card-provider';
-import { MemoryStorage } from '../providers/memory-storage';
-import MainPage from './main-page';
+import { CardProviderStore } from '../../providers/card-provider';
+import { MemoryStorage } from '../../providers/memory-storage';
+import { MainPage } from './main-page';
 
 // Create a mock for the CardProviderStore class
-jest.mock('../providers/card-provider', () => {
+jest.mock('../../providers/card-provider', () => {
   return {
     CardProviderStore: jest.fn().mockImplementation(() => {
       return {
